Updated README.md for Disconnect/Reconnect feature + Updated Disconnect/Reconnect feature
Now there is 4 buttons involving this feature, 2 for disconnecting in a specific fashion and 2 for reconnecting in the same fashion. One of the disconnect buttons will disable RTT and then one of the reconnect buttons only RTT specific features. The other set will disconnect everything (Relay, RTT, and wipe user authentication) and the reconnect will re-initialize, re-authenticate and then reconnect to RTT & Relay. README is updated to instruct how to set this up from the portal side.