Many Small Improvements
The tip of the arrow now aligns with the tip of the user's mouse. Usernames now have a black outline. Gave paint spatters a fade out animation if they have finite lifespan. Added a mask to the game area so the paint does not spill out. Changed labels to refer to paint instead of shockwaves.