Nick/fishnet plugin (#192)
* Added fishnet plugin project

* Added missing files for FishNet

* Cleanup Refactor, compartmentalize Transport Layer

Re-integrated Fishnet from the package manager

* Start Scene - Loads BCManager Singleton, and Added Exit Button, not fully fleshed out

* Lets allow these assets to be synced.

* QUick Find Lobby works now

Still not able to go into a match and out yet.

* Cleanup error on create

* Go into a game come out and find another over and over again

* Init connected clients,

clear them on initialize

* Added display for server host and local player

* Disabled automatic colour changing

* remove sosme warnings debug as new profiles

* cleanup the connect and net id states

* some minor cleanup of things

* Shutdown the server when they close down.

* Migrate owners

* Lets get the transfering of the lobby closer , it looks like the states of objects need to be synced

* Remove the spawn button

Disconnect happens quicker now.

* have the clients send a signal to connect, so the servers can remote connect the

* Remote magic gone

* rename refactor

* refactor and rename

* Sync the player colours when clients reconnect

* Updated to lastest fishnet,

Able to select removeing the client and then reconnecting to the same session using the fishnet HUD. Server when they disconnect their client or their server, should force a migrate host to anothe ruser.

At th emoment it will launch and create new items. The player list items dont seem to respect who previously owned it, or update their sharing mechanisms.
When joining with mutlilpe clients, migrating host seems not as reliable.

* Stopping the connection, then re-establishing a new one

Offers a new netid for the connected server , It removes the previous connected networked objects and, we need to re-create them afterwards

* Lets take away all the complicated stuff for joining and unjoining

Looks like we were just a bit far away since the client managers were not disconnecting. To come, reffreshing the world

* Lets sync up the crown, to denote who is the server.

* When MIgrating Hosts

Cache and reapply changes to clients whom are loaded up between hosts

* Instead of local text, lets highlight your cell such that it displays correctly

* Lets try with four guys, and some minor updates

* Cleanup the BCFishNetTransport file handle the error to go back to the main menu to try and reconnect

* Better build settings for desktop

* Updated some meta data for the builds as well as added a readme

* Lets add a persistent paint blob - updated the readme

* Added a mask and making the splats smaller

* A better persistent calling of the data and map on migrating owned data to those still within the migration

Its still not perfect, but its not terrible

* Join IN progress and a cache contrainer for the main place for the splats and cursors

* migration and join in progres works stable.

It may be doubleing the amount of the items that are spawned during migrations, but will look into it more

* Added the authenticate screen with auto login if possible

Storing the main display name and pushing it into the lobby and other screens

* https://bitheads.atlassian.net/browse/BCLOUD-10943

Able to send signals and relay information based on the user logged in

* auto typing :O

* Lobby is sharing its data with the game state which will be put in progres and try to align with some of the other screens

* Login Cleanup with version text.

* Some Cleanup of the Main Menu

* Refactor the main menu screen

* Update and cleanup op messages and display messages

* Changes for the lobby!

* Update colouring of the login panel

* Refactor the color selector, update the game state to have the colour selector in the game to match the lobby

* Lobby items, in game updates, minor fixes

* Further UI Modifications for loading screen and timer display

* Update BCFishnet.md

* Display Account and Display Name in the main menu

* Updated to brainCloud 5.7

* Removed Login Modal,

Made sure to go back to the main menu, fishnet sync

* Build settings

* Just log the warning

It kicked me back, and when I put networking back on it handled connection again

* Update BCManager.cs

* updated gitignore pattern to include all fishnet items

* these werent saved either

* MAIN!

not main menu

* lets sync these up. again.

* switch back to 5.5

* Point to develop

* A whole slew of UI updates

* lets make them unordered

* Minor UI updates

* Clear messages when host clears canvas

* Better handling of the message clearing

* FIX IT ALL UP!

* Refactor paint splat for better on join and syncing of items

* Some more bugs

version 0.0.12 - When I reQuickjoined the same game, i didnt get joined correctly

bcFishNet - If the player Quickjoins, they are given the default color, instead of their last selected color in this appsession. The default color is black.

* updated version

* Lobby colour selectors

* UI updates

* Multiple players without display name showing Guest_# but some of players are listed in players panel

* Update is host display

* canvas scaler setup

* oops

* a bunch of game changes

* A few other UI updates

Login screen Tab Actions, better random colour choices from the colour palete, update the broken legend in game

* ui name changes for

display name in the lobby and otherwise when the name is blank

* paint splat rotation and scale.

and cleanup

* optmize the paint splats, allow rotations and scalce

* fix the colour selector in the lobby

* update player cursor size on update, update the sprite as well

* better splats

* Cleanup duplicate splats

* MIGRATE MANY

* huzaah

* Some better handling of these events on migration

---------

Co-authored-by: Nick H <nickh@bitheads.com>