Initial Checkin for BCKarton (#209)
* Initial BCKart, Downloaded, installed brainCloud

Added a few helper files BCManager and the like to help streamline bringing kart and bc together seamlessly.

Making notes in notes.md which are very raw which will lead to more details regarding work to integration

* Force Login, and enter in user name

Which is tied to brainCloud for universalId Authentication, and updating the display name prior to logging in.

* Ironedout the Lgoin updating username flow

* Refactored quite a bit

Added a BCLobbyManager to help wiith talking with lobby instance and the new flow.

Host Game, with 8 players is set at the moment.

* Adding Host, Create or Join, and join with error flow

* Update notes.md