Unity Examples Develop Sweeps (#214)
* WIP on updating Auth example

* Updating packages

* Authentication: Updating PackageManager, added Tracking on iOS for Facebook

* Authentication: Update IOSPostBuild script

* Authentication: Adjust Cocoapods setting

* Game Center authentication

* Turn Steam into an external login

* Updating READMEs

* README adjustments

* Adding extra insurances to README

* Adding Jenkin Build to TicTacToe

* Correcting build name for tic tac toe

* One more try with TicTacToe

* Disabled Compression for WebGL builds for TicTacToe

* Changing Tic Tac Toe Project Settings

Change is to set the Enable Exceptions option to Full Without Stacktrace in hopes to resolve WebGL error.

* Bcfn/updateui (#206)

* Initial checkin

* Updated login and lobby screens

Also imported gradient script and UI sprites for backgrounds, buttons and icons

* Ui Changes almost complete

Still need to update Game screen player list items then circle back to polish items

* Finished Updating UI for final screen - Main

Also investigating an issue with changing player color in lobby - WIP

* Completed Ui Update with Known Issue

Known Issue: Player List Item in Lobby has wrong color background target

* Updates and glue for new items

---------

Co-authored-by: markd <markd@bitheads.com>

* BcFishNet - LOBBY > While host exit the lobby, next player does not show as host with crow icon

minor update fo bc fishnet example to update the UI when the lobby member changes

* Updated with Collaborative Canvas logo and brainCloud text added above version

See title of commit

* update project settings and the name of it

* Authentication: Added AuthenticationOculus example

* Authentication: Display error in-editor, fix error display

* - Update Unity Version, add README

* AuthenticationOculus: README fixes

* AuthenticationOculus: README Typo

* Update main README.md
- Also rename BCFishNet README to show properly on GitHub

* Touch up README description

* Update Unity Examples away from the Security Vulnerability  (#208)

* Update BCFishnets package as well as Security Vulnerability for Unity.

* Update TicTacToe

Also ensure that it just auto assigns team members to lobbies, to find more matches. Without needing a server.

* Update RelayTestApp

* Marketplace and Invaders to newest

* bombers version

* bc clashers

* BC Chat

* Auth Example

* AuthenticationOculus: Clean up settings

* Initial Checkin for BCKarton (#209)

* Initial BCKart, Downloaded, installed brainCloud

Added a few helper files BCManager and the like to help streamline bringing kart and bc together seamlessly.

Making notes in notes.md which are very raw which will lead to more details regarding work to integration

* Force Login, and enter in user name

Which is tied to brainCloud for universalId Authentication, and updating the display name prior to logging in.

* Ironedout the Lgoin updating username flow

* Refactored quite a bit

Added a BCLobbyManager to help wiith talking with lobby instance and the new flow.

Host Game, with 8 players is set at the moment.

* Adding Host, Create or Join, and join with error flow

* Update notes.md

* BC Kart Needed More Files (#210)

* We need to track these so others can open it. They were in the ignore list prior

* Added TMP and updated some labels as well as add the version hud to the app lower left corner

* Refactor lobby (#211)

* Starting with ripping out Photon Lobby

Using our own lobby types and replacing what signals were there

* REfacotring lobby,

Not fully working

* REfactoring Lobby signals, And starting the game off of those events with photon

* Launch the track once everyone was connected via photon for the lobby.

* Fix up the lobby handling durign creation

remove the sesion id, since we have our lobbyId that we inject to it

* Removing GameId from GameManager. Updated GameUI to have the lap display info

* Lets call cloud scripts to keep the lobby session alive during long running sessions

* Only send the connected signal the first time connected, otherwise assumptions

are out of order

* Cleanup the way BCKart manages its display, and end of game continuing

The Host will tell everyone the lobbyId and if its theirs they will leave the photon session and go back to the lobby together.

* Refactor how we leave lobbies etc. right now host termination is a lobby termination, while in game

* Modified Race back to 3 Laps

A Sprint to one lap, to allow quicker reuse of lobbies and photon session testing.

* Updated Readme in bckart

* Updated readme to have new sections for the new demos

* ammendment

* Update the LobbyUI on launch -

When a host drops mid game, drop all the other clients back to the lobby scene so that they could restart.

* BCLOUD 11668 add i3D support to cursor party example (#212)

* Added support for i3D in Cursor Party

* Updating Failure Callback for connecting to Relay

* Fusion Karts - Quick Find only finds lobbies at selection of 8. not all the other types.

* Cleanup Kart Login flow, so its seamless

Try forcing player count to 8 in each game

* updated build version

* Lobby Display count,

Error displayed when leaving the photon session as the host and going back to the lobby.
Disable hud interactions and show feedback that the game is starting

* Dont show the host left message when continuing

* Strip Max Players from lobby

search only by lobbies of 8

* Disconnect and launch smoothly, hide starting on disable UI

* added kart selection display UI in the main hud, auth error when logging in with bad info

* Updating display in lobby of who's you, use signals to sync up the kart Id across clients.

* Bug fix was to rearrange the log in process abit (#213)

The bug would occur because it would get stuck on a request completing from client side

* Update Jenkins Build to NOT overwrite the target dir if its already set via command line

* Remove Join Option

After all racers have crossed the finish line, then "Waiting for players" is replaced with "Race Finished". Wait 2 seconds, then automagically return to the lobby.

---------

Co-authored-by: Michael Costa <michaelc@bitheads.com>
Co-authored-by: Michael Costa <68234970+michaelc-bitheads@users.noreply.github.com>
Co-authored-by: franco <82105876+francobithead@users.noreply.github.com>
Co-authored-by: markd <markd@bitheads.com>